refactor: migrate Popover component to use react-aria-components and remove v2 dependency
